requirements.txt
# Discord Bot Requirements
# Główna biblioteka Discord
discord.py>=2.3.0

# Obsługa zmiennych środowiskowych
python-dotenv==1.0.1

# HTTP client do komunikacji z API
aiohttp>=3.8.0

# Panel administracyjny
Flask>=2.3.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
bcrypt>=4.0.1
PyJWT>=2.7.0
psutil==5.9.8

# Dodatkowe zależności (opcjonalne, ale zalecane)
# Jeśli planujesz rozbudowę bota:
requests==2.31.0
asyncio>=3.4.3
werkzeug==3.0.1
